log the username which was authenticated, not first principal
In case of multiple principals in a certificate, cert.ValidPrincipals[0]
will create  a log with first principal name which has no relation with the
authentication. It creates misleading logs. It assumes first principal
will always be the username, which might not be the case.

Instead, the username should be logged explicitly that was authenticated.

Signed-off-by: Shivam Dixit <shivamd001@gmail.com>